+++ c/include/searches.h
@@ -1,12 +1,14 @@
 #ifndef _SEARCHES_H_
 #define _SEARCHES_H_
 
+#include <stddef.h>
+
 typedef struct SearchResult {
-  unsigned int index;
+  size_t index;
   char found;
 } SearchResult;
 
-SearchResult binary_search(int* arr, unsigned int start, unsigned int end,
+SearchResult binary_search(int* arr, size_t start, size_t end,
                            int x);
 
 #endif  // _SEARCHES_H_
