+++ c/src/search/binary_search.c
@@ -1,6 +1,6 @@
 #include "searches.h"
 
-SearchResult binary_search(int* arr, unsigned int start, unsigned int end,
+SearchResult binary_search(int* arr, size_t start, size_t end,
                            int x) {
   SearchResult r;
   if (start <= end) {
